DeleteUserServlet, delUser.jsp - 회원 탈퇴 오류 수정 NaverAddInfoServlet, NaverLoginFrmServlet, AddNaverInfo.jsp, NaverLoginService, NaverLoginDao- 네이버 로그인 관련 수정 myPageHeader.jsp - msg 추가 pwChk.jsp - swal추가, joinValidate(조건문구) 추가

+++ src/main/java/kr/or/iei/user/controller/DeleteUserServlet.java
@@ -10,8 +10,11 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
+import org.mindrot.jbcrypt.BCrypt;
+
 import kr.or.iei.common.JDBCTemplate;
 import kr.or.iei.user.model.service.UserService;
+import kr.or.iei.user.model.vo.UserSite;
 
 /**
  * Servlet implementation class DeleteUserServlet
@@ -33,28 +36,50 @@ public class DeleteUserServlet extends HttpServlet {
 	 */
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		String userNo = request.getParameter("userNo");
+		String pwChk = request.getParameter("pwChk");
+		System.out.println(userNo);
+		System.out.println(pwChk);
+		HttpSession session = request.getSession();		
 		
-		UserService service = new UserService();
-		int result = service.deleteUser(userNo);
-		
-		if(result > 0) {
-			HttpSession session = request.getSession(false);
+		if(session != null) {
+			UserSite loginUser = (UserSite)session.getAttribute("loginUser");
 			
-			if(session != null) {
-				session.invalidate();
+/*			//사이트 회원인 경우 Bcrypt로 기존 비밀번호와 비교해서 일치하지 않으면 return으로 종료
+			if(!BCrypt.checkpw(pwChk, loginUser.getUserPw())) {
+				request.setAttribute("title", "알림");
+				request.setAttribute("text", "기존 비밀번호가 일치하지 않습니다");
+				request.setAttribute("icon", "error");
+				request.setAttribute("loc", "/user/delUserFrm");
+
+				return;
+			}else if(loginUser.getUserType() != 3){
+				//네이버 회원이 아닌경우 return 으로 종료
+				request.setAttribute("title", "알림");
+				request.setAttribute("text", "");
+				request.setAttribute("icon", "error");
+				request.setAttribute("loc", "/user/delUserFrm");
+				
+				return;
 			}
+*/
+			//비크립트로 비교한 pw값이 true or 유저타입이 3인경우
+		if(BCrypt.checkpw(pwChk, loginUser.getUserPw()) == true || loginUser.getUserType() == 3) {
+			UserService service = new UserService();
+			int result = service.deleteUser(userNo);
 			
-			request.setAttribute("title", "알림");
-			request.setAttribute("msg", "회원 탈퇴가 완료되었습니다");
-			request.setAttribute("icon", "success");
-			request.setAttribute("loc", "javascript:window.opener.location.href=\"/\"; self.close();");		
-		}else {
-			request.setAttribute("title", "알림");
-			request.setAttribute("msg", "회원 탈퇴 중 오류가 발생했습니다");
-			request.setAttribute("icon", "error");
-			request.setAttribute("loc", "javascript:self.close()");
+				if(result > 0) {
+					session = request.getSession(false);
+					
+					if(session != null) {
+						session.invalidate();
+					}
+					response.getWriter().print("1");		
+				}else {
+					response.getWriter().print("-1");
+				}
+				request.getRequestDispatcher("/WEB-INF/views/common/msg.jsp").forward(request, response);
+			}
 		}
-		request.getRequestDispatcher("/WEB-INF/views/common/msg.jsp").forward(request, response);
 	}
 
 	/**
